change some names in app.js change some css style in app.css change authentication files move admin files  - from /pages to /pages/admin/  -from /components to /components/admin
