+++ src/frontend/webappService/src/components/admin/RoleForm.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import Button from './Button';
+import Button from '../Button';
 import '../css/RoleForm.css';
 
 function RoleForm() {
