+++ src/frontend/webappService/src/components/admin/UserProfilAdmin.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
-import Button from "../components/Button.js";
-import "../css/App.css";
-import "../css/UserProfil.css";
+import Button from "../Button.js";
+import "../../css/App.css";
+import "../../css/UserProfil.css";
 import { useNavigate } from "react-router-dom";
 
 function UserProfilAdmin({ title, userId }) {
